shell: set the password via the MYSQL_PWD enviroment variable
This improves further on the idea of getting rid of the ugly warning
(https://github.com/planetscale/cli/pull/99) when you pass the password
via the flag.

Instead of saving the password to a tmp file, we're setting it as a
temporary environment variable inside the Go process. The environment
variable can be still read from outside (via `ps`), but @piki mentioned
that both are worse in terms of security and on the same level, so it
should be OK to use this approach.