proxy: pick up a random port and bind to the correct port
This PR changes the address to bind locally. Instead of binding to an
hardcoded port, we now let the OS pick up a random port for us. In order
to make this work, we had to introduce a new method to the `sql-proxy`
repo (https://github.com/planetscale/sql-proxy/pull/52).

Both `pscale shell` and `pscale connect` now automatically binds to a
random port.

closes https://github.com/planetscale/project-big-bang/issues/111