cmd: remove `org` subcommand
The `org` subcommand is empty and not used right now. We also don't konw
how exactly we're going to use. Second, the `org switch` subcommand is
alread exposed as `switch`, so there is little use to it.

To avoid confusing, we should remove the `org` subcommand. We should
later rethink on how we want to use it.