cmd: show user friendly error messages
Most of the time we print the output of `cmd.Usage()` if an argument is
missing. However that's not very useful. It's like dumping out the
content of a book. Instead we should tell the user what exactly went
wrong. Even better, we can aid/help them (such as let them pick up a
branch name), but that's an improvement for another PR.

This PR removes all `cmd.Usage()` calls and replaces them with explicit
and actionable error messages. We use a custom `RequiredArgs` function
that checks whether the given arguments exist for the given subcommand.
As a bonus, we don't have to check for arguments anymore and it
simplifies our CLI quite a bit.

Before:

```
$ pscale deploy list planetscale
Usage:
  pscale deploy list [database] [branch] [flags]

Flags:
  -h, --help   help for list

Global Flags:
      --api-token string   The API token to use for authenticating against the PlanetScale API.
      --api-url string     The base URL for the PlanetScale API. (default "https://api.planetscaledb.io/")
      --config string      config file (default is $HOME/.config/planetscale/config.yaml)
      --json               Show output as JSON
      --org string         The organization for the current user
```

After:

```
$ pscale deploy list
Error: missing arguments <database, branch>

Usage: pscale deploy list <database> <branch> [flags]
```

If argument exist but we require multiple, it only shows the ones that
are missing:

```
$ pscale deploy list planetscale
Error: missing argument <branch>

Usage: pscale deploy list <database> <branch> [flags]
```