cmd: fix outputting resources for CVS mode if only one exists.
This PR fixes the `--format csv` output for commands that only output a
single item. The `gocsv` package is not able to print individual structs
and expects a slice of structs. We fix the issue by expliclity wrapping
individual items with a slice.

There is a side effect though. This changes how we print in JSON mode:
instead of printing a single item it's now printing a list with a single
item in it.

fixes: https://github.com/planetscale/project-big-bang/issues/224