cli: check for latest version and inform the user
This PR introduces a new check whenever a user makes an API Call. To prevent spamming the GitHub API and also stop nagging the user, we only check once for every 24 hours. The user can disable version checking with the `PSCALE_NO_UPDATE_NOTIFIER` as well.

The business logic is inspired from GitHub's `gh` tool, but it's of course adapted to our needs.

fixes: https://github.com/planetscale/project-big-bang/issues/216