dumper: don't create the schemas
Vitess doesn't support `CREATE DATABASE` statements. This means that the upcoming `pscale database restore` command fails as it's not able to create them. In this PR we're removing the ability to create the schema creation files.

There is work to support it in Vitess (https://github.com/planetscale/project-big-bang/issues/287#issuecomment-832153649), but the scope is out of beta. Because we're also owning the "database creation" as part of our product, we should remove it.

closes: https://github.com/planetscale/project-big-bang/issues/287