Add licensed CI check
This PR adds a new CI workflow that checks continously for new
dependencies and their licenses. We use the
https://github.com/github/licensed tool for this check.

`licensed` caches all the dependencies licenses and then checks
according to the `.licensed.yml` file whether a license is allowed or
not. This will make sure we detect dependencies we introduce have the
correct licenses and agreements:

```
allowed:
  - mit
  - apache-2.0
  - bsd-2-clause
  - bsd-3-clause
  - mpl-2.0
```

/xref https://github.com/planetscale/project-big-bang/issues/307