connect: introduce `--host` and `--port` flags
This PR introduces two new flags to choose the local host and port.  Instead of using the `--local-addr` flag, we let users select the port and host in a more easy way.

Also, it looks like, we would print the wrong address even if the user has choosed a custom local address.