tty: detect tty without depending on the color package
This fixes the issue when coloring is disabled with `NO_COLOR` enviroment variable. This would set `color.NoColor` to true and hence in an enviroment where `NO_COLOR` is set, commands like `pscale auth login` would fail:

```
$ pscale auth login
Error: The 'login' command requires an interactive shell
```

Now, if the user sets the `NO_COLOR` environment variable, it still continues to work, but only the color's are disabled.