cmd: improve the authentication warning for service token's
Not all users want to use access tokens while using the CLI. Some users prefer to create a service token and use that. However, our warning assumes that pscale was called interactively, and the user intends to use an access token.

This PR improves the message to make it clear that there are other options.