Clarify success message of deploy command
`pscale deploy-request deploy <db> <deploy-request-id>` actually adds the deploy-request to the queue to be deployed, but the current success message tells you that the deploy-request has *been* deployed already.

This modifies the success message for clarity. Ideally, it'd also print the URL to view the deploy request on the Planetscale Dashboard.