Add M1 Mac homebrew install fallback link
I have been struggling to install the CLI, especially through Homebrew. I found a StackOverflow that explains the steps to be able to install it through Rosetta which was the only way I could get it to work. Thought it would be worth adding a little footnote to the step as a fallback if someone else also runs into this.