cmd: show a warning for self compiled binaries
Let's make sure users don't use an compiled version accidentally. For us devs, it doesn't matter because we can set the environment variable so everything would be the same as previously. For people who have compiled it themselves wouldn't realize later that they use an old version. If `pscale` is compiled from source, it also can't do a version check, so we don't have a way to prompt the user for an update.