ping: map up endpoints to display names
The awkward part about this is because we need to "resolve" what the
optimized endpoint mapped up to.

The "optimized" endpoints are kinda synthetic regions, so we can only
determine what they point to by asking the endpoint.

We surface the information about the endpoint unofficially through the
Server HTTP header on a response, so we can determine what we're
pointing at by making a HEAD request to the optimized endpoint and
parsing the Server header.

Short of adding a real "info" or something API to our edge layer, this
is the best and most supported way and isn't intending to change in the
nera or mid-term future. If we ever do explose some explicit `/info`
endpoint (which we may), this can be changed to accommodate.