cmd/password/list: show TTL remaining and if password is expired
This adds two new columns to the password list, "TTL REMAINING", and
"EXPIRED".

This allows a user to see, if they have a credential that is ephemeral,
to see how long it has left before it's expired, and if it is expired
already.

Relies on https://github.com/planetscale/planetscale-go/pull/193