Disable custom signal handling if not on foreground
If pscale is used in scripts with the explicit
PSCALE_ALLOW_NONINTERACTIVE_SHELL env flag, disable the signal changes
we do for making the shell work in interactive mode.

This check also needs to verify that all of stdin, stdout & stderr are a
proper TTY and not just stdout, since if you pipe in only input for
example we still need to disable this since there's no foregrounding.