Code cleanup plus additional flags
BuildKite was failing due to the way I was passing the helper used for printing so I made some adjustments and simplified how that worked.

Afterward I added in two more optional flags: `--schema-only` and `--data-only`.

If both are used at the same time then that's essentially the equivalent of a normal "restore" where both are being included.

The new `--start-from` parameter can also be used together with these too.