Adjusted --start-from to --starting-table and --ending-table
I was thinking about this some more and it makes sense to allow for this option to be more comprehensive and allow not just >= a starting table to be included but a <= an ending table too or being able to provide a more limited starting/ending table range so I made some adjustments.

This removes the `--start-from` option and replaces it with the `--starting-table` and `--ending-table` options which can be used separately or together.