Fix the MySQL version detection and only use up to MySQL 8.x
MySQL 9.x is backwards compatibility breaking as it removes support for
`mysql_native_password` which is what is used with PlanetScale as the
default at the moment.

So enforce MySQL 8.x as the latest at the moment.

Also upgrade planetscale-go and add some empty test mocks for now so it
can compile. This update removes a bunch of legacy code so it can't be
used accidentally either.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>