Added better support for views within dumper.go
Currently, `dumper.go` is able to export view definitions without issue, but it unfortunately also runs the view and includes, or attempts to include, the data associated with the view too which is unnecessary.

The changes here add support for collecting the list of views so that processing can be slightly adjusted if a "table" is actually a view.

In that situation, the same `-schema-view.sql` file suffix is utilized that the MyDumper project uses for views and the step that unnecessarily runs the view query is skipped to avoid that issue.