Added context to longer-running functions for quicker cancellation
Currently, when using either the `pscale database dump` or `pscale database restore-dump` commands there may be times where you would like to cancel out of the operation because it is taking too long, a mistake was made, etc.

At the moment the behavior for both commands is not ideal in these cases since you can attempt to send a `Ctrl+C` command but it won't be very responsive / immediate as you expect it to be.

Instead, depending on the length of time the operation has remaining, it may actually continue to completion and/or the renewal of the temporary 5 minute credential won't occur and it will stop / error then.

There is likely a more comprehensive approach to using contexts properly here that my limited Go knowledge is unaware of, but it seemed like passing in the current context into the longer-running functions that do the bulk of the data exporting / loading was enough to make `Ctrl+C` responsive enough for my expectations.