Added show details output information for views
Now that https://github.com/planetscale/cli/pull/932 was merged in a little earlier I was able to incorporate some extra show details output specific for views into this PR now.

The current `canRestoreSchema()` has also been wrapped around the portion of code that restores the view definitions. In the future we may want to differentiate between restoring tables vs. views only but for now they are both considered to be part of the schema so I'm using the same check for both at the moment which should work ok for now.