Improved context usage and added extra error output
While running throught the commands again today I noticed that if an error was encountered while a query was executing, which could happen when using the new `--data-only` flag if a prior attempt had been canceled partway through, that the subsequent data files would continue to process.

This could be helpful perhaps in some cases but in most we would probably want the restore attempt to error out early in a case like this so that the issue can be addressed and some useful error information displayed.

This commit adds some improved context usage so that things will error out earlier and also adds a snippet of the larger error output we would normally return that way the user might be able to see the error more easily and then adjust their usage of the command or manually empty the table before trying again.