Removed IsAuthenticated() check
This variation allows for new code within the `NewClientFromConfig()` method to automatically detect / correct if the service token values have been provided in a swapped manner and adjusts things so things can continue working.

This bit of code does include one line to the console that will indicate that the values have been swapped which I think is a good idea to at least provide the user with some feedback that the values have been passed in incorrectly.