Suppress Authorization header after redirects
Requests like `pscale api
/organizations/$org/databases/$db/branches/$branch/query-patterns` get
redirected to an S3 bearer URL.  `pscale api` follows those redirects.  It
needs to not include the `Authorization` header when talking to S3,
because including that header is a security leak, and also causes S3 to
reject the request.

The fix is to teach the OAuth2 http client to not follow redirects that
go to a new root domain (e.g., somewhere other than planetscale.com).
Instead of following the redirect automatically in the OAuth2 http client,
now we construct a new, clean http client without the Authorization
header.

This fixes a security leak, in which a client's OAuth2 bearer token would
be forwarded to S3.  It also fixes a bug, in which `.../query-patterns`
requests just didn't work at all.