Implement `pscale shell` for Pg
--replica and --role work as expected.  --local-addr is refused because
psql connects directly, not through a local proxy.  --remote-addr is
supported but unlikely to be useful.

We hard-code port :5432 instead of going through the connection pooler on
:6432, because the pooler pools connections by logged-in role.  We're
using an ephemeral role, so we'd always get assigned to our own
connection.  We'd be paying the extra hop through the pooler for no
benefit.