WIP - make `pscale shell` work with Pg
This is an attempt to make `pscale shell` work with Postgres databases.
It works as far as identifying that a branch is Postgres and finding a
copy of `psql` or `psql-17` to shell out to.

Unfortunately, the passwords API endpoint doesn't work with Pg yet, so we
pass `psql` a bunch of incorrect parameters, and it hangs trying to
connect to the wrong cloud endpoint.

We could push this a bit further using the undocumented Roles API, but
there's still no way to create or retrieve a valid password yet.