Fix prepare binary data
This updates to a Vitess branch still on v21 where we now have a fix
that ensures we don't mark all values as binary, which breaks bind vars.

This fix is also present on v22 already, but we can't update to that yet
until we have the full and entire fleet on at least v22.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>